requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
asttokens==3.0.0
attrs==25.4.0
autopep8==2.3.2
aws-requests-auth==0.4.3
beautifulsoup4==4.14.2
bedrock-agentcore==0.1.5
bedrock-agentcore-starter-toolkit==0.1.14
boto3==1.40.45
botocore==1.40.45
Brotli==1.1.0
certifi==2025.10.5
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.4.3
click==8.3.0
comm==0.2.3
ddgs==9.6.0
debugpy==1.8.17
decorator==5.2.1
diagram==0.2.28
diagrams==0.24.4
dill==0.4.0
distlib==0.4.0
docstring_parser==0.17.0
duckduckgo_search==8.1.1
executing==2.2.1
filelock==3.20.0
frozenlist==1.8.0
graphviz==0.20.3
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
hyperframe==6.1.0
identify==2.6.15
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.0
ipykernel==6.30.1
ipython==9.6.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
jsonschema==4.25.1
jsonschema-path==0.3.4
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.8.1
lazy-object-proxy==1.12.0
lxml==6.0.2
markdown-it-py==4.0.0
markdownify==1.2.0
MarkupSafe==3.0.3
matplotlib-inline==0.1.7
mcp==1.16.0
mdurl==0.1.2
moviepy==2.2.1
mpmath==1.3.0
multidict==6.7.0
nest-asyncio==1.6.0
nodeenv==1.9.1
numpy==2.3.3
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.2
opentelemetry-api==1.37.0
opentelemetry-instrumentation==0.58b0
opentelemetry-instrumentation-threading==0.58b0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
packaging==25.0
parso==0.8.5
pathable==0.4.4
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.4.0
prance==25.4.8.0
pre_commit==4.3.0
primp==0.15.0
proglog==0.1.12
prompt_toolkit==3.0.52
propcache==0.4.0
psutil==7.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-openapi-schema-to-json-schema==0.0.3
pycodestyle==2.14.0
pydantic==2.11.10
pydantic-settings==2.11.0
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.3
pyzmq==27.1.0
questionary==2.1.1
referencing==0.36.2
requests==2.32.5
rfc3339-validator==0.1.4
rich==14.1.0
rpds-py==0.27.1
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.14
s3transfer==0.14.0
shellingham==1.5.4
six==1.17.0
slack_bolt==1.25.0
slack_sdk==3.36.0
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.8
sse-starlette==3.0.2
stack-data==0.6.3
starlette==0.48.0
strands-agents==1.10.0
strands-agents-tools==0.2.9
sympy==1.14.0
tenacity==9.1.2
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typer==0.19.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uv==0.8.23
uvicorn==0.37.0
virtualenv==20.34.0
watchdog==6.0.0
wcwidth==0.2.14
wrapt==1.17.3
yarl==1.22.0
zipp==3.23.0
